Add files via upload
This is my 4th semester computer networks project. It is a QR code generator using socket programming in python. Once the server is run and connected to local host , the client is run and the user can enter a url and a qr code is generated for the given url. We have used python libraries like socketserver, http.server, urllib.parse, BytesIO for server and QRcode, tkinter, requests, PIL, IO for client.